Test Coordinator.heartbeat, refactor stream test helpers #17
